(maint) Install CA certs during git and package runs
Previously, the step for installing CA certs was only run during "git"
acceptance runs. This step is needed for module related acceptance tests
to work on platforms that don't already trust the forge's root
certificate, notably on Windows.

This commit makes the setup step common for both git and packages. This
doesn't actually change anything for "packages" because the step is only
executed on Windows, and we don't yet test against packages on Windows.
It also makes it easier to add support for other agent platforms that
have the same issue, e.g. Solaris

This commit also fixes the misspelling in the file name.