(PUP-3274) Update doc for doc application to reflect removed features
There is an example that implies you can point the doc tool at a
directory and it will do its thing. This feature was actually for
pointing at specific manifest files and would call manifestdoc() which
now raises an error (since feature is removed). Everything works fine
(for the level of fine currently associated with the rdoc tool :) if you
instead set modulepath.

This removes the text that describes that a single manifest can be
scanned for documentation. It changes the now faulty example to be a
modulepath example.