(PUP-2080) Add function to the puppet language
This adds the ability to write functions in the puppet language.
(Resurrected from earlier version on master to comply with the
changes made for 4.0).

Puppet functions have qualified names. They are loaded from the
functions directory relative to the module or environment.
They can also be defined in manifests directly and are then directly
instantated and given to the respective loader. It is an error to
overload an already loaded function, but if the function is defined via
a manifest first, it will shadow the function in /functions.