(maint) Do not add delivery domain to generated hosts
...unless it is missing. Earlier versions of Beaker, I believe, provided
a host name that was not an fqdn, and when our Rakefile was creating a
config to allow Beaker to be re-run against preserved hosts, it would
add the delivery domain. Now, hostnames are being found fully qualified
in the logs, and we need to only add the domain if the hostname is not
qualified.

This commit also refactors the code that releases vms to make use of the
internal vmpooler rather than relying on Beaker::VcloudPooled, whose
interface may change.