(PUP-3016) use puppet helper in acceptance tests
Use the puppet() helper function instead of assuming puppet is in the
path. After installing puppet on a Windows system in particular, it's
path is not reflected in the shell environment. Using puppet() lets
Beaker figure out how to execute the puppet binary.
Retain less-than argument for version of module installed.

Get around having to quote the whole command by inserting a space
after the version operator.  sigh.

In our current windows-cygwin overloaded ENV one has to single quote the entire
'puppet module #args <options>' command, which is not currently possible
with the helper (thus the use of Command in the previous version of this test).

The only other way i could get it to work is by quoting the version
string, but this requires a space after the operator. Otherwise
windows/cygwin says "file not found".

closes #3342