(maint) Add validation tests for top-level constructs in conditionals
In PUP-3274 some tests ended up "missing" (or at least hard to see
where certain behavior was asserted). This brings back three such
tests for testing that top-level constructs class, define and node
can not be defined inside of a conditional expression.

The tests are added to the validator_spec since it is responsible for
flagging these as errors.