(PUP-405) Remove acceptance test that requires active_record
This removes the acceptance test "ENC node information is used when
store configs enabled (#16698)" since it tests the combination of
an ENC and the use of active record based store configs.

A corresponding test should be added to test a real supported
combination of services (node classifier / puppet db), but this cannot
be done in the puppet code base / acceptance suite. See PUP-3707 for
follow up.