(PUP-3754) Better error message when failing to chown log file
Previously, puppet would log an error if it failed to chown a
file-based log destination, but didn't specify what owner/group it tried
to chown to or the path to the file, making it difficult to debug.

This commit changes the error message to include the owner, group and
file path for easier debugging.