(PUP-2351) Add a puppet epp command for validation and rendering
This adds a new command called puppet-epp that can validate
one or several epp templates, render them to text with
various options for specifying facts and variables in scope,
giving variables to the template etc.

The command also has a dump action that is useful for debugging
epp parsing. This command is similar to the puppet parser dump
action.

For more details run 'puppet man epp'