(PUP-3645) Add support for configurable URL prefix
This commit adds support for a configurable setting `master_url_prefix', which
defaults to '/puppet', that the master's http endpoints are mounted at. All
requests generated by the agent honor this prefix. This will allow us to mount
the Puppet HTTP endpoints in a namespaced URL space alongside other apps
running on the same webserver.