(PUP-3839) Inherit Gem mirror in acceptance tests
The testing infrastucture should be resilient to network issues. For
acceptance tests that exercise the Gem provider and setup that installs
Gems, there hasn't been a clean way to specify that the internal mirror
should be used. This follows the Gemfile pattern and selects the
mirror if GEM_SOURCE is specified.

OUTSIDE_PUPPETLABS has also been removed, as we decided it was a bad
practice to default to something that only works on a private network.
A notice was added to ci-pe-puppet.sh to ensure sqa-utils is added
to a Gemfile.local when running the script manually.