(PUP-3268) Provide default module path dir for fallback production
Legacy environments were guaranteed to produce an environment instance
for any random environment name, and supply it with a modulepath, which
had the side effect of ensuring that there would always be a target_dir
for the module_tool.  Directory environments raise an error if the
environment does not exist, but they have an edge case in that the
default production environment is always made available for
bootstrapping. In the case of a call like `puppet config` this allows it
to function even when Puppet is misconfigured. But for module_tool, this
provides an environment for production which has no modulepath, and now
causes an obscure error.

In continuation with the theme that Puppet should continue to just work
in the naive unconfigured case, this commit adds the
'$environmentpath/production' path to fallback production environment
modulepath so that the module_tool will continue to have a target_dir in
this case.