(maint) Prefer agent -t
For the agent_runs_pluginsync_with_proper_environment test, changing the
agent to run with "-t" includes the --no-daemonize and --ontime flags,
but also ensures --verbose is set which provides better debugging
output.  This also means --detailed-exitcodes is set, but exit code for
thisi operation is expected to be the default 0. The run_agent_on()
Beaker helper is deprecated, and this switches to using on(agent,
puppet()) instead.

Changes to use_enc_environment_for_pluginsync are just to use a single
call of agent.puppet['vardir'] (and thus a single puppet config call on
the agent).