(PUP-3268) Do not attempt to set environmentpath to an empty string
With legacy environments removed, environmentpath is no longer an
optional setting.  This test is checking behavior when environmentpath
is set in the master section of puppet.conf, and was setting
environmentpath in 'main' to the empty string.  This configuration is no
longer possible, and the main section has been removed.