(PUP-3268) Pass the actual environment instance in file_content requests
Resource::Catalog.environment() is the String representation of the
environment. We were passing this when the File type content parameter
was set to lookup file contents through the indirectory (via
puppet:///something), and the indirector request would implicitly lookup
an Environment instance based on the String. With legacy environments
removed, however, an agent during pluginsync (or handling any other
remote file content calls during catalog application) is now configured
with directory environments which will raise an error if it cannot find
an actual directory environment matching the environment string.

Since the agent's concept of environment is meant to be entirely
symbolic, it should not be necessary to configure these environments on
the agent. Instead, this commit passes
Resource::Catalog.environment_instance() which is an actual (remote)
Puppet::Node::Environment that the indirector will not need to look up.