(PUP-1072) deal with HTTP errors
The indirector terminus just returns nil if the HTTP request fails. This
accomodates the usual file serving semantics. If the user specified
multiple sources, the agent can fall back to the next source.

Therefor, an error should not be raised. Added a debug message to allow
the user to see what's going on.