(PUP-1085) Add better handling of pacman package groups.
This patch constructs a virtual version for package groups, consisting
of lines in the format:
<pkgname> <pkgver>

This allows detection wether a package in the group is out of date.
For example for the resource...

package { 'gnome':
    ensure => latest,
}

...puppet will produce the following output (note the epiphany package
version upgrade):

Notice: /Stage[main]/Main/Package[gnome]/ensure: current_value
baobab 3.12.1-1
empathy 3.12.4-1
eog 3.12.2-1
epiphany 3.12.1-1
[...snip...]
, should be
baobab 3.12.1-1
empathy 3.12.4-1
eog 3.12.2-1
epiphany 3.12.1-2
[...snip...]

Also, package groups that are partially installed are considered absent
now. This prevents the provider to think that the group `gnome` is
installed, just because the package `eog` is installed, which was
previously the case.