(PUP-1085) Fix uninstall of pacman package groups.
Currently pacman gets executed without the --recursive option when
uninstalling a group.  This prevents the uninstalling of package groups
because it needs to be done recursively.  The fix is to pass the -s
option to remove when the resource is a group.

This also fixes uninstall to use yaourt if present, so that puppet can
be run as a normal user (assuming yaourt can password-less sudo).