Quote file modes as recommended. Also other strings.
Add some quotes around strings that occur in block comments.  These
block comments are converted into HTML documentation that is published
on the official Puppet documentation Web pages at Puppetlabs.  It is
important that the examples be correct, because Puppet users are very
likely to uses these examples as a reference.

I noticed this because my colleagues started putting quotes around file
modes (e.g. mode => '0644').  From looking at the examples, I thought
the correct way to write these were as numbers.  But when I looked up
the documentation on the "mode" attribute of the "file" resource type,
it clearly says:

    This value should be specified as a quoted string; do not use
    un-quoted numbers to represent file modes.

So my colleagues seem to be right, but the examples are wrong.  This is
what the commit is trying to fix.

I also noticed that in one of the examples, the string values of the
"owner" and "group" attributes weren't quoted.  This seems to be an
error as well, so I fixed that too.

Another small change to the examples is that the numeric file modes are
written with four digits, i.e. with leading zeroes.