(maint) Allow git URI to be specified in local_options.rb
Previously, the Rakefile assumed if you were doing a `git` type
acceptance test that the repos to install were of the form
`puppet#stable` or just `puppet`. In either case, environment variables
GIT_SERVER, FORK, and SHA were used to generate the actual URI, e.g.
git://<server>/<fork>#<sha>

However, if the repo to install was an actual URI, then the Rakefile
would silently map each repo to `nil`, and would trigger the error
message added in 099b0cbd6.

This scenario would always happen if you ran the tests once, and copied
`merged_options.rb` to `local_options.rb`, and attempted to re-run the
tests with the same configuration as last time. In that case,
`local_options.rb` would contain URIs of the form:

   :install=>
     ["git://github.com/puppetlabs/facter.git#stable",
      "git://github.com/puppetlabs/hiera.git#stable",
      "git://github.com/puppetlabs/puppet.git#master"
     ],

This commit modifies the Rakefile to preserve the specified URI if it
doesn't match one of our "short" versions, e.g. puppet#stable. This only
applies to `git` type tests.