(maint) Remove brittle test for storeconfigs enabled
This test is now failing on platforms whose native ruby is 1.8.7,
because it installs activerecord, which depends on i18n, which was
updated on Dec 19th 2014 to require ruby 1.9.3. So it would take
some work to make this test work on any of those platforms.

For context, this test is historically super brittle (look at the commit
history to watch how it's been confined to fewer and fewer platforms),
and it uses deprecated functionality, now removed in puppet 4.

So this commit simply removes the test, so that we can spend our
time on more value-adding efforts.