(PUP-3912) Remove must alias for rspec 3
Previously, puppet had a monkey patch to rspec to alias 'must'
to 'should'. This was done to workaround the use of a 'should'
method in Puppet::Type conflicting with rspec's should.

With rspec 3 syntax, we can replace those with expect(...).to
syntax and do away with the monkey patch. (An earlier commit
programatically replaced all the actual 'should's in the spec
tests using transpec, but it missed the 'must's.)

This commit:
* remove the monkey patch to rspec
* uses the expect(...).to syntax everywhere there was a .must
* replaces use of the == method with eq
* includes some other minor rspec-3 changes to files that needed
  the above treatment