(PUP-3912) Make some methods public where specs were relying on them
Previously these methods were made private, but spec tests were still
using them as if they were public. Apparently, rspec 2 didn't enforce
access methods but rspec 3 does, so for the time being simply mark
the methods public (with a note to the user that they shouldn't be
treated as part of a public API).

As a possible future improvement the code could be rewritten to not
test using private methods, but that looks like a fair bit of surgery
for not that much gain at this time. Something to consider as/when
this code is refactored going forward.