(PUP-3645) Add not_found helper
Puppet's route processing will throw a 'Not Found' error if the first part of
a route has no matches, but will consider a route successfully matched if the
first part matches, even if the rest does not.

With the addition of a url prefix to the puppet master api, we now have
`/v2.0` and `/v3` chained underneath `/puppet`. This means that a request to
`/unknown` would cause a 'Not Found' error to be raised, but a request to
`/puppet/unknown` would result in an empty response.

This commit adds a not_found helper, which can be included anywhere we want to
explicitly specify that a 404 should be raised.  In this case, it is chained
under the puppet master prefix, after `/v2.0` and `/v3`.