(PUP-3912) Rewrite expect to use: not_to raise_error
Previously these two tests had an expect for a nil return (which
the tests don't care about) and no expect for the lack of a raise
(which was the point of the test). So rewrite to be explicit
about what is expected, which also makes the tests in the 'describe'
block more obviously parallel.

This diff best seen with -w.