(PUP-3912) Replace conditional pending blocks
Rspec 3 drops support for conditional pending blocks, so
replace them.

Since a rspec 2 'pending' is an rspec 3 'skip' replace some with:
  if condition
    skip
  else
    <test>
  end

Or in some cases, just replace with:
  describe "stuff", :if => condition
    <test>
  end
This isn't strictly a pending or a skip, but this was used
in cases where the condition at hand (a particularly ruby bug)
was out of the hands of this code base, so it's not really
*pending* anything, it's just a dependency.

This also exposed a couple 'should' -> 'expect ... to' instances
not converted by transpec so I fixed those.